ui/flask_ui/requirements.txt
flask
simplejson
Pillow
